fix: round 반환형 및 금융기관목록 조회하는 핸들러 return 값 수정

+++ src/main/java/com/kakao/demo/service/dto/SupportedAmountOfInstitution.java
@@ -9,9 +9,9 @@ public class SupportedAmountOfInstitution {
     private SupportedAmountOfInstitution() {
     }
 
-    public SupportedAmountOfInstitution(int year, int amount) {
+    public SupportedAmountOfInstitution(int year, double amount) {
         this.year = year;
-        this.amount = amount;
+        this.amount = Double.valueOf(amount).intValue();
     }
 
     public int getYear() {
